Heroc: Dex pre-optimization (ODEX)
Our processor is slow. First boot after flashing non-optimized nightly
builds takes forever, especially when flashing over an existing install.

Pre-optimized builds actually take up less space on our cramped device
(~104MB in /system versus ~91MB in /system and ~25MB in dalvik-cache).

And the update ZIP itself is only ~1MB larger with pre-optimization,
so it's not much of a bandwidth hit if that was of concern.

Change-Id: Id4c283f69ca2690f80ca4d9d3c242ce2e21972d2