heroC: Fix for mms recieve
mms recieve on heroC is not working due to changes
in  content disposition parsing in PduParser.java.
This change fixes the if statement which previously ran
if the preference config_mms_content_disposition_support
was true, which it always was.
This change sets the preference in the heroC
overlay config.xml to be false.

Change-Id: Ib27a0cba5e2d8c4423b5cf433161c6f0ed1d1dc7